loadFilterFromString
